// management/frontend/src/api/schemas/businessLineCreate.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Ego Collection Management
 * OpenAPI spec version: 0.1.0
 */

export interface BusinessLineCreate {
  /**
     * @minLength 1
     * @maxLength 128
     */
  name: string;
  external_id?: string | null;
  parent_id?: string | null;
  path?: string | null;
  status?: number;
}