// management/frontend/src/api/zod/workspace-business-lines/workspace-business-lines.ts
export const workspaceBusinessLinesCreateBusinessLineBodyNameMax = 128;

export const workspaceBusinessLinesCreateBusinessLineBodyExternalIdOneMax = 128;

export const workspaceBusinessLinesCreateBusinessLineBodyParentIdOneMax = 128;

export const workspaceBusinessLinesCreateBusinessLineBodyPathOneMax = 512;

export const workspaceBusinessLinesCreateBusinessLineBodyStatusDefault = 1;

export const WorkspaceBusinessLinesCreateBusinessLineBody = zod.object({
  "name": zod.string().min(1).max(workspaceBusinessLinesCreateBusinessLineBodyNameMax),
  "external_id": zod.union([zod.string().max(workspaceBusinessLinesCreateBusinessLineBodyExternalIdOneMax),zod.null()]).optional(),
  "parent_id": zod.union([zod.string().max(workspaceBusinessLinesCreateBusinessLineBodyParentIdOneMax),zod.null()]).optional(),
  "path": zod.union([zod.string().max(workspaceBusinessLinesCreateBusinessLineBodyPathOneMax),zod.null()]).optional(),
  "status": zod.number().default(workspaceBusinessLinesCreateBusinessLineBodyStatusDefault)
})
// ...
export const workspaceBusinessLinesUpdateBusinessLineBodyNameMax = 128;

export const workspaceBusinessLinesUpdateBusinessLineBodyExternalIdOneMax = 128;

export const workspaceBusinessLinesUpdateBusinessLineBodyParentIdOneMax = 128;

export const workspaceBusinessLinesUpdateBusinessLineBodyPathOneMax = 512;

export const workspaceBusinessLinesUpdateBusinessLineBodyStatusDefault = 1;

export const WorkspaceBusinessLinesUpdateBusinessLineBody = zod.object({
  "name": zod.string().min(1).max(workspaceBusinessLinesUpdateBusinessLineBodyNameMax),
  "external_id": zod.union([zod.string().max(workspaceBusinessLinesUpdateBusinessLineBodyExternalIdOneMax),zod.null()]).optional(),
  "parent_id": zod.union([zod.string().max(workspaceBusinessLinesUpdateBusinessLineBodyParentIdOneMax),zod.null()]).optional(),
  "path": zod.union([zod.string().max(workspaceBusinessLinesUpdateBusinessLineBodyPathOneMax),zod.null()]).optional(),
  "status": zod.number().default(workspaceBusinessLinesUpdateBusinessLineBodyStatusDefault)
})